btrfs: don't background writeout incomplete pages
Summary:
Log files are usually opened with O_APPEND and then slowly written to.
Because btrfs requires stable pages we could background writeout the last page
that hasn't been fully filed, and so whoever is logging could experience a
latency spike when trying to write() to the logfile.  Fix this by just skipping
the last page if it isn't completely full for background writeouts.

This commit squashes together a series of related patches in this area,
which makes it easier to port around and test.

Btrfs: limit delalloc to full pages in append mode

Summary:
We made it so that writepage wouldn't start IO on the last partial page
but forgot to make delalloc not actually fill in the last partial page.  This is
the other half of the fix to make O_APPEND files not get blocked waiting on IO
for the last partial page of a file.

Tasks: 5549520

Btrfs: clear O_APPEND hack bit when closing file

We set the BTRFS_INODE_APPEND_WRITE flag on an inode whenever we write
to it using an O_APPEND file descriptor. However, we never clear this
flag, which means that the tail pages hang around forever until a
WB_SYNC_ALL comes along. Fix it by clearing the flag when the file is
released. This can prematurely clear the flag if the file is referenced
by multiple O_APPEND file descriptors, but that's not super common in
practice, and the flag will be set again the next time the other file
descriptor writes to it, anyways.

Btrfs: force write out O_APPEND tail page when freeing space

We hit an early ENOSPC bug caused by our O_APPEND hack preventing pages
from being written out when flushing to free space. It is trivial to
reproduce:

    # for ((i = 0; i < 100000; i++)); do dd
    if=/dev/zero of=/mnt/test/$i oflag=append count=1 status=none || break;
    done
    dd: failed to open ‘/mnt/test/44679’: No space left on device
    # touch /mnt/test/foo
    touch: cannot touch ‘/mnt/test/foo’: No space left on device
    # sync
    # touch /mnt/test/foo

Fix it by writing out the tail page if we are doing writeback to free
space. The flush_space() path already sets the reason to
WB_REASON_FS_FREE_SPACE (through btrfs_writeback_inodes_sb_nr()).

Btrfs: fix infinite loop in O_APPEND optimization

As part of the O_APPEND optimization, find_lock_delalloc_range() will
truncate the delalloc range to not include a partial page past i_size.
However, if this partial page is the only page we are writing out, then
delalloc_end will be less than delalloc_start, and we'll get stuck in an
ininite loop.

Usually, __extent_writepage() skips writepage_delalloc() entirely if
we're only writing out the tail page for an O_APPEND file, so we don't
hit this path. However, it's possible for the BTRFS_INODE_APPEND_WRITE
flag to be set after __extent_writepage() has already checked it and
before find_lock_delalloc_range() checks it, which triggers the bug.
This window became more common after 10e7e796c413 ("Btrfs: clear
O_APPEND hack bit when closing file"). Previously, this bit would only
go from unset to set on the first O_APPEND write to an inode. Now that
we clear the bit on close, it can happen any time the inode is reopened
O_APPEND and written to.

The fix is to only look at the bit once and pass it around.

Btrfs: don't reread i_size for O_APPEND optimization

There appears to be a similar issue to the one I fixed in commit
a0a4959ecc64 ("Btrfs: fix infinite loop in O_APPEND optimization"),
which was caused by rereading BTRFS_INODE_APPEND_WRITE. If the file is
truncated between __extent_writepage() and find_lock_delalloc_range(),
we can end up in a similar situation where we shrink delalloc_end down
to the new i_size and end up with delalloc_end < delalloc_start. I only
found this through code reading, I didn't reproduce it.

Btrfs: use O_APPEND hack for all appending writes

Some applications, including RocksDB, do sequential, extending writes to
files opened without O_APPEND and thus hit the same latency issues with
stable pages that we attempted to solve with our hack. Instead of gating
this on O_APPEND, just detect writes starting at EOF, and always clear
the flag on closing a file descriptor.

Signed-off-by: Chris Mason <clm@fb.com>