block: ensure that trim size doesn't exceed UINT_MAX
If the user settable max trim size exceeds the amount of bytes we
can store in an unsigned int, then we can create trims that exceed
the size of a request. If the trim is exactly 4GiB in size, then
it'll appear like a zero sized trim. If it's larger, it won't trim
as much as it should.

Fixes: 23010cbacad5 ("block: add BLKDEV_DISCARD_SYNC flag")
Signed-off-by: Jens Axboe <axboe@fb.com>
(cherry picked from commit 8006cd1249f497ecbb14290116aa29def1de9803)