tpm_tis: work around HP G10 TPM bug
The TPM in the HP G10 machines seems to have a bug which causes the
kernel to believe that the TPM is ready to receive a command when it
actually isn't. This causes the send to time out in get_burstcount().
This bug is demonstrated by the userspace reproducer in P64781264. We
can apparently work around it by reading from the status register one
extra time.

Signed-off-by: Omar Sandoval <osandov@fb.com>
(cherry picked from commit b1f0d561e825b567d872a7d1a64232a4861b271b)