mm: memcontrol: memory.high.tmp, timeout-based memory.high setting
The senpai process can kick the bucket during a memory consumption dip
(crashes; ooms; gets killed), in which case the workload will struggle
when it picks up again, with potentially nobody around to remedy that.

As a failsafe, introduce a timeout-based memory.high.tmp setting that
is only in effect for a specified time. Senpai can use this to safely
probe for lower memory footprints. Once expired, whatever was set in
memory.high goes back into effect.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>