blk-iocost: Provide extra stats for IO cost model verification
Summary:

Andy wants the following extra information so that he can verify
correctness of the cost model.

* Rand IO proportions.
* Latency miss ratios.

Provide those informatino by exposing the followings.

* rmet, wmet, rmissed, wmissed and rqwait in the root cgroup.
* rrandbytes, wrandbytes, rrandios, wrandios in all cgroups.

blkcg stat code leaves a lot to be desired and can't currenlty be
easily extended to support per-controller hierarchical stats, so the
rand io counters are added to the main rwstat counters and used only
for iocost controller.  It's not the prettiest but should do the job
for now.  In the longer term and for upstream, we should move blkcg to
use rstat.

Test Plan: manually tested

Reviewers: axboe, jbacik

Subscribers: newella

Tasks:

Tags:
(cherry picked from commit 22ba2721d33892264d038ddaadedf6aacca47758)