net: fb-only: dctcp fallback to cubic in the absence of ECN
Summary:
net: fb-only: dctcp fallback to cubic in the absence of ECN

This is an internal patch for fb so that dctcp will fallback
to cubic instead of reno when the ECN support is missing.
Otherwise, throughput performance loss is observed after
falling back to reno.

The main purpose is to remove one of many variables when
production-izing dctcp.

This patch will be dropped once the migration to cgroup-v2 is
completed because bpf@cgroup-v2 has a way to select congestion
control.

It requires tcp_cubic to be built-in (instead of a module). It then does
a deep copy of the "cubictcp" with some clean-up on the "key", "list"
and "name".  The "name" is changed to "dctcp-cubic" to make it clear
in the "ss" output:

ESTAB    0        0               10.0.0.2:8000         10.0.0.1:57360 [...]
dctcp-cubic wscale:7,7 rto:201 rtt:0.448/0.224 mss:1448 [...]

Test Plan:

Signed-off-by: Martin KaFai Lau <kafai@fb.com>
(cherry picked from commit 9872651bea3d14160eba68eba951f7ebb4406575)
(cherry picked from commit f5f6672181c46dc2f56469bcfef574f831ce06eb)
Signed-off-by: Kyle McMartin <jkkm@fb.com>