mm: page_counter: mitigate consequences of a page_counter underflow
When the unsigned page_counter underflows, even just by a few pages, a
cgroup will not be able to run anything afterwards and trigger the OOM
killer in a loop. Underflows shouldn't happen, but when they do in
practice, we may just be off by a small amount that doesn't interfere
with the normal operation - consequences don't need to be that dire.

Reset the page_counter to 0 upon underflow. We'll issue a warning that
the accounting will be off and then try to keep limping along.

Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
(cherry picked from commit a8bd081cf2b868211a6ae1d245302dd99e8d1afa)