fs: inode: Reduce volatile inode wraparound risk when ino_t is 64 bit
Summary:
fs: inode: Reduce volatile inode wraparound risk when ino_t is 64 bit

In Facebook production we are seeing heavy inode number wraparounds on
tmpfs. On affected tiers, in excess of 10% of hosts show multiple files
with different content and the same inode number, with some servers even
having as many as 150 duplicated inode numbers with differing file
content.

This causes actual, tangible problems in production. For example, we
have complaints from those working on remote caches that their
application is reporting cache corruptions because it uses (device,
inodenum) to establish the identity of a particular cache object, but
because it's not unique any more, the application refuses to continue
and reports cache corruption. Even worse, sometimes applications may not
even detect the corruption but may continue anyway, causing phantom and
hard to debug behaviour.

In general, userspace applications expect that (device, inodenum) should
be enough to be uniquely point to one inode, which seems fair enough.
This patch changes get_next_ino to use up to min(sizeof(ino_t), 8) bytes
to reduce the likelihood of wraparound. On architectures with 32-bit
ino_t the problem is, at least, not made any worse than it is right now.

I noted the concern in the comment above about 32-bit applications on a
64-bit kernel with 32-bit wide ino_t in userspace, as documented by Jeff
in the commit message for 866b04fc, but these applications are going to
get EOVERFLOW on filesystems with non-volatile inode numbers anyway,
since those will likely be 64-bit. Concerns about that seem slimmer
compared to the disadvantages this presents for known, real users of
this functionality on platforms with a 64-bit ino_t.

Other approaches I've considered:

- Use an IDA. If this is a problem for users with 32-bit ino_t as well,
this seems a feasible approach. For now this change is non-intrusive
enough, though, and doesn't make the situation any worse for them than
present at least.
- Look for other approaches in userspace. I think this is less
feasible -- users do need to have a way to reliably determine inode
identity, and the risk of wraparound with a 2^32-sized counter is
pretty high, quite clearly manifesting in production for workloads
which make heavy use of tmpfs.

Test Plan:
See merge commit at top of stack

Differential Revision: https://phabricator.intern.facebook.com/D19498647

Signed-off-by: Chris Down <chris@chrisdown.name>
(cherry picked from commit e4497399d049bb21ee97d688f31d0fd13d75033d)